fix: PR 리뷰 피드백 반영 (세기 추론 로직 개선, 유틸리티 분리, 신규 필드 CSV 매칭 추가)
