+++ src/admin/wards-management/dto/create-ward.dto.ts
@@ -1,5 +1,6 @@
 import { Transform } from 'class-transformer';
 import {
+  ArrayNotEmpty,
   IsArray,
   IsDateString,
   IsEmail,
@@ -8,30 +9,7 @@ import {
   IsString,
 } from 'class-validator';
 import { IsValidPhone } from '../validators/phone-number.validator';
-
-/**
- * Parse birthdate from YYMMDD or YYYY-MM-DD format
- * Century inference: 00-30 → 2000s, 31-99 → 1900s
- */
-function parseBirthDate(value: string | undefined): string | undefined {
-  if (!value || typeof value !== 'string') return undefined;
-  const trimmed = value.trim();
-  if (trimmed.length === 0) return undefined;
-
-  // Already in ISO format (YYYY-MM-DD)
-  if (/^\d{4}-\d{2}-\d{2}$/.test(trimmed)) return trimmed;
-
-  // YYMMDD format
-  if (/^\d{6}$/.test(trimmed)) {
-    const yy = parseInt(trimmed.substring(0, 2), 10);
-    const mm = trimmed.substring(2, 4);
-    const dd = trimmed.substring(4, 6);
-    const century = yy <= 30 ? '20' : '19';
-    return `${century}${trimmed.substring(0, 2)}-${mm}-${dd}`;
-  }
-
-  return trimmed; // passthrough for validation
-}
+import { parseBirthDate, sanitizeDiseases } from '../../../common/utils/date.utils';
 
 export class CreateWardDto {
   @IsString()
@@ -80,6 +58,7 @@ export class CreateWardDto {
   @IsOptional()
   @IsArray()
   @IsString({ each: true })
+  @Transform(({ value }) => sanitizeDiseases(value))
   diseases?: string[];
 
   @IsOptional()
@@ -109,3 +88,4 @@ export class CreateWardDto {
   )
   notes?: string;
 }
+
